fix: run store single-threaded to avoid sqlite deadlock
sqlite: turn on WAL journal mode by default as this is better for multiple
processes it seems.

Note: I've left in the multithreaded code under an option config.multithread,
which is defaulted to False and not set through the config file currently. In
case we every want to revisit it. But it does look like it's not really
supported by sqlite.