add query command for ad hoc reporting
Note: the query command allows you to specify multiple sql files to run, or
otherwise takes the query text from stdin. If you have a multiline comment
as the first line of the query, it uses that as a title for the output.

main: change common spec parameter to optional

   - config_file adapter: parse default section, which is used if no spec
     parameter is passed.